chore: tsc strict, CI test matrix
