+++ .github/workflows/lint-test-sdk.yml
@@ -21,28 +21,34 @@ on:
 env:
   SDK_BRANCH_NAME: ${{ inputs.sdk_branch  || github.head_ref || github.ref_name || 'main' }}
   TEST_DATA_BRANCH_NAME: ${{ inputs.test_data_branch || 'main' }}
-        
+
 jobs:
   lint-test-sdk:
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        node-version: [ '18', '20', '22', '23' ]
     steps:
       - name: Display Testing Details
         run: |
           echo "Running SDK Test using"
           echo "Test Data: sdk-test-data@${TEST_DATA_BRANCH_NAME}"
           echo "SDK Branch: js-client-sdk@${SDK_BRANCH_NAME}"
-      - uses: actions/checkout@v2
-        with:
-          repository: Eppo-exp/js-client-sdk
-          ref: ${{ env.SDK_BRANCH_NAME }}
+      - uses: actions/checkout@v3
       - name: Use Node.js
-        uses: actions/setup-node@v1
+        uses: actions/setup-node@v3
         with:
-          node-version: '18.x'
-      - uses: actions/cache@v2
+          node-version: ${{ matrix.node-version }}
+      - name: Get yarn cache directory path
+        id: yarn-cache-dir-path
+        run: echo "dir=$(yarn cache dir)" >> $GITHUB_OUTPUT
+      - uses: actions/cache@v4
+        id: yarn-cache # use this to check for `cache-hit` (`steps.yarn-cache.outputs.cache-hit != 'true'`)
         with:
-          path: './node_modules'
-          key: ${{ runner.os }}-root-node-modules-${{ hashFiles('./yarn.lock') }}
+          path: ${{ steps.yarn-cache-dir-path.outputs.dir }}
+          key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
+          restore-keys: |
+            ${{ runner.os }}-yarn-
       - name: Install SDK dependencies
         run: yarn --frozen-lockfile
         working-directory: ./
@@ -54,19 +60,25 @@ jobs:
         working-directory: ./
   typecheck:
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        node-version: [ '18', '20', '22', '23' ]
     steps:
-      - uses: actions/checkout@v2
-        with:
-          repository: Eppo-exp/js-client-sdk
-          ref: ${{ env.SDK_BRANCH_NAME }}
+      - uses: actions/checkout@v3
       - name: Use Node.js
-        uses: actions/setup-node@v1
+        uses: actions/setup-node@v3
         with:
-          node-version: '18.x'
-      - uses: actions/cache@v2
+          node-version: ${{ matrix.node-version }}
+      - name: Get yarn cache directory path
+        id: yarn-cache-dir-path
+        run: echo "dir=$(yarn cache dir)" >> $GITHUB_OUTPUT
+      - uses: actions/cache@v4
+        id: yarn-cache # use this to check for `cache-hit` (`steps.yarn-cache.outputs.cache-hit != 'true'`)
         with:
-          path: './node_modules'
-          key: ${{ runner.os }}-root-node-modules-${{ hashFiles('./yarn.lock') }}
+          path: ${{ steps.yarn-cache-dir-path.outputs.dir }}
+          key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
+          restore-keys: |
+            ${{ runner.os }}-yarn-
       - name: Install SDK dependencies
         run: yarn --frozen-lockfile
         working-directory: ./
