+++ src/index.ts
@@ -515,7 +515,7 @@ export async function init(config: IClientConfig): Promise<EppoClient> {
       // both failed, make the "fatal" error the fetch one
       initializationError = initFromFetchError;
     }
-  } catch (error) {
+  } catch (error: any) {
     initializationError = error;
   }
