+++ tsconfig.json
@@ -8,7 +8,7 @@
     "target": "es2017",
     "outDir": "dist",
     "noImplicitAny": true,
-    "strictNullChecks": true
+    "strict": true
   },
   "include": [
     "src/**/*.ts"
